// src/shared/schemas/account-api.ts
export interface components {
	schemas: {
// ...
		Account_Locations: {
			/**
			 * @description CRM: Account.name
			 * @example iba2d
			 */
			readonly name?: string;
			/**
			 * @description CRM: Account.display_name_c
			 * @example IBA IT Park - 2nd Dep
			 */
			readonly display_name?: string;
			/**
			 * @description CRM: Account.dep_id_c
			 * @example 2d
			 */
			dep_id?: string;
			/**
			 * @description CRM: Account.dep_name_c
			 * @example 2nd Department
			 */
			dep_name?: string;
			/**
			 * @description CRM: Account.org_id_c
			 * @example iba
			 */
			org_id?: string;
			/**
			 * @description CRM: Account.org_name_c
			 * @example IBA IT Park
			 */
			org_name?: string;
			/** @description CRM: Location objects related to Account */
			locations?: components["schemas"]["Account_Location"][];
		};
// ...
		Location: {
			/**
			 * @description CRM: icdc_location.name
			 * @example xby
			 */
			readonly name?: string;
			/**
			 * @description CRM: icdc_location.display_name
			 * @example OneCloud, Sofia
			 */
			display_name?: string;
			/**
			 * Format: url
			 * @description CRM: icdc_location.url
			 * @example https://api.xby.icdc.io
			 */
			readonly url?: string;
			/**
			 * Format: url
			 * @description CRM: icdc_location.service_provider
			 * @example stone
			 */
			readonly service_provider?: string;
			/**
			 * Format: url
			 * @description CRM: icdc_location.back_to_url; Section 'Home Page' in UI
			 * @example https://ibacloud.by
			 */
			back_to_url?: string;
			/**
			 * Format: url
			 * @description CRM: icdc_location.status_page_url; Section 'Home Page' in UI
			 * @example https://status.ibacloud.by
			 */
			status_page_url?: string;
			/**
			 * Format: url
			 * @description CRM: icdc_location.privacy_policy_url; Section 'Home Page' in UI
			 * @example https://ibacloud.by/privacy-policy-en
			 */
			privacy_policy_url?: string;
			/**
			 * Format: url
			 * @description CRM: icdc_location.cookies_policy_url; Section 'Home Page' in UI
			 * @example https://ibacloud.by/cookies-policy-en
			 */
			cookies_policy_url?: string;
			/**
			 * Format: url
			 * @description CRM: icdc_location.tos_url; Section 'Home Page' in UI
			 * @example https://ibacloud.by/terms-of-service-en
			 */
			tos_url?: string;
			/**
			 * Format: url
			 * @description CRM: icdc_location.dpa_url; Section 'Home Page' in UI
			 * @example https://ibacloud.by/data-processing-aggrement-en
			 */
			dpa_url?: string;
			/**
			 * @description CRM: icdc_location.footer_message; Section 'Home Page' in UI
			 * @example ®2020 IBA IT Park. All rights reserved.
			 */
			footer_message?: string;
		};
// ...
	};
// ...
}